attempt1/requirements.txt
numpy==1.18.4
Pillow==7.1.2
pkg-resources==0.0.0
scipy==1.4.1
